requirements.txt
comtypes
python-pptx
pandas
regex
openpyxl